refactor: extract widget routes into routes_widgets.py blueprint
Move /widget_config/<widget_name> out of app.py into a new
routes_widgets.py blueprint named "widgets". Behavior unchanged.

The captured `app.logger` reference inside the handler becomes
`current_app.logger` for blueprint-friendliness. No template
url_for() calls reference this endpoint, so no template changes.
